relayer/relayer/requirements.txt
djangorestframework==3.13.1
